feat: Put all the application logic in a separate package
