--- server/apps/accounts/implemented/__init__.py
@@ -1,1 +1,0 @@
-from .signup import signup_implementation # Noqa
