+++ server/apps/accounts/logic/implemented.py
@@ -1,8 +1,9 @@
 from punq import Container
-from server.apps.accounts.helpers import validate_password_format
-from server.apps.accounts.interfaces import PasswordValidator, UserCredentialsRepository
-from server.apps.accounts.repositories import DatabaseRepository
-from server.apps.accounts.services import SignUpService
+
+from .helpers import validate_password_format
+from .interfaces import PasswordValidator, UserCredentialsRepository
+from .repositories import DatabaseRepository
+from .services import SignUpService
 
 
 signup_implementation = Container()
