+++ tests/test_apps/test_accounts/test_logic/test_services/test_signup_service.py
@@ -1,8 +1,8 @@
 # Todo: refactor these tests
 from pytest import fixture, mark, raises
-from server.apps.accounts.dto_types import RawFormData, RawUserCredentials
-from server.apps.accounts.exceptions import AccountAlreadyExists, MismatchedPasswords, UncorrectPassword
-from server.apps.accounts.services import SignUpService
+from server.apps.accounts.logic.dto_types import RawFormData, RawUserCredentials
+from server.apps.accounts.logic.exceptions import AccountAlreadyExists, MismatchedPasswords, UncorrectPassword
+from server.apps.accounts.logic.services import SignUpService
 
 
 class CredentialsMockRepository:
